+++ R/edl_set_token.R
@@ -1,7 +1,3 @@
-# Proto-package functions for dealing with NASA's earthdata login
-# Requires GDAL >= 3.6
-
-
 #' Get or set an earthdata login token
 #'
 #' This function will ping the EarthData API for any available tokens.
@@ -12,9 +8,10 @@
 #' popular R packages which use gdal to immediately authenticate any http
 #' addresses to NASA EarthData assets.
 #'
-#' IMPORTANT: it is necessary to unset this token using `edl_unset_token()`
-#' **before** trying to access HTTP resources that are not part of EarthData,
-#' as setting this token will cause those calls to fail!
+#' IMPORTANT: it is necessary to unset this token using [edl_unset_token()]
+#' before trying to access HTTP resources that are not part of EarthData,
+#' as setting this token will cause those calls to fail! OR simply use
+#' [edl_netrc()] to authenticate without facing this issue.
 #'
 #' NOTE: Because GDAL >= 3.6.1 is required to recognize the GDAL_HTTP_HEADERS,
 #' but all versions recognize GDAL_HTTP_HEADER_FILE. So we set the Bearer token
@@ -28,17 +25,26 @@
 #' environmental variable?  logical, default TRUE.
 #' @param format One of "token", "header" or "file."  "header" adds the prefix
 #' used by http headers to the return string.  "file" returns
+#' @param prompt_for_netrc Often netrc is preferable, so this function will by
+#' default prompt the user to switch.  Set to FALSE to silence this.
 #' @return A text string containing only the token (format=token),
 #' or a token with the header prefix included, `Authorization: Bearer <token>`
 #' @export
 #' @examplesIf interactive()
 #' edl_set_token()
+#' edl_unset_token()
 edl_set_token <- function (username = default("user"),
                            password = default("password"),
                            token_number = 1,
                            set_env_var = TRUE,
-                           format = c("token", "header", "file")
+                           format = c("token", "header", "file"),
+                           prompt_for_netrc = interactive()
 ){
+
+  if(prompt_for_netrc){
+    done <- we_prefer_netrc(username, password)
+    if (done) return(invisible(TRUE))
+  }
 
   p <- edl_api("/api/users/tokens", username, password)
 
@@ -61,6 +67,8 @@ edl_set_token <- function (username = default("user"),
   invisible(out)
 }
 
+
+
 edl_headerfile <- function(token) {
   header = edl_header(token)
   headerfile <- tempfile(pattern="GDAL_HTTP_HEADERS", fileext = "")
@@ -74,11 +82,6 @@ edl_setenv <- function(token) {
   Sys.setenv("GDAL_HTTP_HEADER_FILE"=headerfile)
 }
 
-default <- function(what) {
-  switch(what,
-         user = Sys.getenv("EARTHDATA_USER", "earthaccess"),
-         password = Sys.getenv("EARTHDATA_PASSWORD", "EDL_test1"))
-}
 
 
 edl_header <- function(token) {
@@ -102,83 +105,6 @@ edl_api <- function(endpoint,
   p
 }
 
-#' download assets from earthdata over https using bearer tokens
-#'
-#' NOTE: This should be used primarily as a fallback mechanism!
-#' EarthData Cloud resources are often best accessed directly over
-#' HTTPS without download.  This allows subsets to be extracted instead
-#' of downloading unnecessary bits.  Unfortunately, certain formats do
-#' not support such HTTP-based range requests (e.g. HDF4), and require
-#' the asset is downloaded to a local POSIX filesystem first.
-#' @param href the https URL of the asset
-#' @param dest local destination
-#' @param header the Authorization header (`Bearer <token>`).  Will be
-#' requested automatically if not provided.
-#' @param use_httr logical, default TRUE. Should we use httr or base method?
-#' @param ... additional arguments to the [utils::download.file()] (ignored
-#' if `use_httr=TRUE`).
-#' @return the `dest` path, invisibly
-#' @export
-#' @examplesIf interactive()
-#' href <- lpdacc_example_url()
-#' edl_download(href)
-edl_download <- function(href,
-                         dest = basename(href),
-                         header = edl_set_token(),
-                         use_httr = TRUE,
-                         ...) {
-  bearer <- paste("Bearer", header)
-  if (use_httr) {
-    httr::GET(href,
-              httr::write_disk(dest, overwrite = TRUE),
-              httr::add_headers(Authorization = bearer))
-  } else {
-    utils::download.file(href, dest, ...,
-                         header = list(Authorization = bearer))
-  }
-  invisible(dest)
-}
-
-
-#' Receive and set temporary AWS Tokens for S3 access
-#'
-#' @param daac the base URL for the DAAC
-#' @param username EarthDataLogin user
-#' @param password EarthDataLogin Password
-#' @return list of access key, secret key, session token and expiration,
-#' invisibly.  Also sets the corresponding AWS environmental variables.
-#'
-#' @examplesIf interactive()
-#' edl_s3_token()
-#' @export
-edl_s3_token <- function(daac = "https://data.lpdaac.earthdatacloud.nasa.gov",
-                         username = default("user"),
-                         password = default("password")) {
-
-  p <- edl_api(endpoint = "/s3credentials",
-               username = username,
-               password = password,
-               base = daac)
-  Sys.setenv(AWS_ACCESS_KEY_ID = p$accessKeyId)
-  Sys.setenv(AWS_SECRET_ACCESS_KEY = p$secretAccessKey)
-  Sys.setenv(AWS_SESSION_TOKEN = p$sessionToken)
-
-  invisible(p)
-}
-
-#' Replace https URLs with S3 URIs
-#' @param href a https URL from an EarthData Cloud address
-#' @param prefix the preferred s3 prefix, e.g. `s3://` (understood by gdalcubes),
-#' or `/vsis3`, for terra/stars/sf or other GDAL-based interfaces.
-#' @return a URI that strips basename and protocol and appends prefix
-#' @export
-#' @examples
-#' href <- lpdacc_example_url()
-#' edl_as_s3(href)
-edl_as_s3 <- function(href, prefix = "s3://") {
-  p <- httr::parse_url(href)
-  paste0(prefix, p$path)
-}
 
 
 #' URL for an example of an LP DAAC COG file
@@ -196,18 +122,6 @@ lpdacc_example_url <- function() {
 }
 
 
-
-#' Helper function for extracting URLs from STAC
-#'
-#' @param items an items list from rstac
-#' @param assets name(s) of assets to extract
-#' @return a vector of hrefs for all discovered assets.
-#'
-edl_stac_urls <- function(items, assets = "data") {
-  purrr::map(items$features, list("assets")) |>
-    purrr::map(list(assets)) |>
-    purrr::map_chr("href")
-}
 
 
 #' unset token
@@ -258,4 +172,21 @@ edl_revoke_token <- function(
 
   }
   invisible(p2)
+}
+
+
+
+
+
+
+#' Helper function for extracting URLs from STAC
+#'
+#' @param items an items list from rstac
+#' @param assets name(s) of assets to extract
+#' @return a vector of hrefs for all discovered assets.
+#'
+edl_stac_urls <- function(items, assets = "data") {
+  purrr::map(items$features, list("assets")) |>
+    purrr::map(list(assets)) |>
+    purrr::map_chr("href")
 }